https://github.com/schnellvishnu/Job.git

+++ seeker/Recruter/views.py
@@ -9,9 +9,12 @@ from Applicant.models import Apply_Job
 class Navbarview(View):
        def get(self,request):
              session_uname=request.session.get('loggedinuser')
-             my_string = ' '.join([str(element) for element in session_uname]) 
-             print("1111111",my_string)                       
-             return render(request,"Navbar.html",{"username": my_string})                  
+        #      my_string = ' '.join([str(element) for element in session_uname]) 
+             print("1111111", session_uname)  
+             if (session_uname == None):
+                     return redirect("login")
+             else:                     
+               return render(request,"Navbar.html",{"username": session_uname})                  
 # class AddCompanyview(CreateView):
                    
 #         model=Compnay_Profile
@@ -19,6 +22,7 @@ class Navbarview(View):
 #         fields="__all__"
 #         success_url = reverse_lazy("employer_jobs")        
 class Addjobview(CreateView):
+                            
         model =Add_Jobs
         template_name="Addjob.html"
         # fields="__all__"
@@ -40,65 +44,72 @@ class GetCompanyPro_ListJobview(View):
         def get(self,request):
               session_uname=request.session.get('loggedinuser')
               print('0000',session_uname) 
-              my_string = ' '.join([str(element) for element in session_uname])
-              print("1111",my_string)
-              s_obj=Signup_User.objects.filter(username=my_string).values()
-              print(s_obj[0]['name'])
-              companyname_var=s_obj[0]['name']
-              
-            
-                          
-              obj=Add_Jobs.objects.filter(company_name=companyname_var)
-              obj2=Compnay_Profile.objects.filter(company_name=companyname_var)
-              print("hhjh",obj[0].application_count)
-              
-            
-              return render(request,"List_job.html",{"jobs":obj,"company":obj2,"username":my_string})  
+              if (session_uname == None):
+                return redirect("login")
+        #       my_string = ' '.join([str(element) for element in session_uname])
+        #       print("1111",my_string)
+              else:
+                s_obj=Signup_User.objects.filter(username= session_uname).values()
+                print(s_obj[0]['name'])
+                companyname_var=s_obj[0]['name']
+                
+                
+                                
+                obj=Add_Jobs.objects.filter(company_name=companyname_var)
+                obj2=Compnay_Profile.objects.filter(company_name=companyname_var)
+                print("hhjh",obj[0].application_count)
+                
+                
+                return render(request,"List_job.html",{"jobs":obj,"company":obj2,"username":session_uname})  
 def filterjob(request):
-        if request.method == 'POST':
-            name = request.POST.get('name')
-            print("name",name)
-            location = request.POST.get('location')
-            date=request.POST.get('date')
-            print(date)
-            loggedin_user=request.session.get('loggedinuser')
-            my_string = ' '.join([str(element) for element in  loggedin_user]) 
-        try:      
-            if(name!=""and location !="" and date!=""):
-                  obj=Add_Jobs.objects.filter(job_name=name).filter(location=location).filter(date=date)
-                  if(obj[0] == []):
-                                                      
-                     obj="No Data Found"   
-            elif(location =="" and date ==""):  
-                  obj=Add_Jobs.objects.filter(job_name=name)
-                  if(obj[0] == []):
-                                                      
-                     obj="No Data Found"   
-            elif(name=="" and date ==""):  
-                  obj=Add_Jobs.objects.filter(location=location)
-                  if(obj[0] == []):
-                                                      
-                     obj="No Data Found"   
-            elif( date ==""):  
-                  obj=Add_Jobs.objects.filter(job_name=name,location=location)
-                  if(obj[0] == []):
-                                  
-                     obj="No Data Found"   
-            elif( name =="" and location ==""):  
-                  obj=Add_Jobs.objects.filter(date=date)
-                  print("jj") 
-                  if(obj[0] == []):
-                     print("jj")                       
-                     obj="No Data Found"  
-            elif(obj[0] == []):
-                  obj="No Data Found"                            
-           
-              
-        except:
-            obj="No Data Found"      
-        print(obj)    
-          
-        return render(request,"List_job.html",{"jobs":obj})
+        loggedin_user=request.session.get('loggedinuser')                    
+        if (loggedin_user == None):
+                return redirect("login")   
+        else:                 
+                if request.method == 'POST':
+                        name = request.POST.get('name')
+                        print("name",name)
+                        location = request.POST.get('location')
+                        date=request.POST.get('date')
+                        print(date)
+                        loggedin_user=request.session.get('loggedinuser')
+                #     my_string = ' '.join([str(element) for element in  loggedin_user]) 
+                try:      
+                        if(name!=""and location !="" and date!=""):
+                                obj=Add_Jobs.objects.filter(job_name=name).filter(location=location).filter(date=date)
+                                if(obj[0] == []):
+                                                                
+                                   obj="No Data Found"   
+                        elif(location =="" and date ==""):  
+                                obj=Add_Jobs.objects.filter(job_name=name)
+                                if(obj[0] == []):
+                                                                
+                                   obj="No Data Found"   
+                        elif(name=="" and date ==""):  
+                                obj=Add_Jobs.objects.filter(location=location)
+                                if(obj[0] == []):
+                                                                
+                                   obj="No Data Found"   
+                        elif( date ==""):  
+                                obj=Add_Jobs.objects.filter(job_name=name,location=location)
+                                if(obj[0] == []):
+                                                
+                                   obj="No Data Found"   
+                        elif( name =="" and location ==""):  
+                                obj=Add_Jobs.objects.filter(date=date)
+                                print("jj") 
+                                if(obj[0] == []):
+                                   print("jj")                       
+                                   obj="No Data Found"  
+                        elif(obj[0] == []):
+                                obj="No Data Found"                            
+                        
+                        
+                except:
+                   obj="No Data Found"      
+                   print(obj)    
+                
+                return render(request,"List_job.html",{"jobs":obj})
                                                
                            
 class Editjobview(UpdateView):
@@ -110,29 +121,40 @@ class Editjobview(UpdateView):
         pk_url_kwarg="id" 
         success_url=reverse_lazy("employer_jobs")
 class Deletejobview(View):
-        def get(self,request,id):             
+        def get(self,request,id):
+                loggedin_user=request.session.get('loggedinuser')                    
+                if (loggedin_user == None):
+                     return redirect("login") 
+                else:           
                     obj=Add_Jobs.objects.get(id=id)
                     obj.delete()
                     return redirect("employer_jobs")
             
        
 def Viewapplication(request,job_code):
+        loggedin_user=request.session.get('loggedinuser')                    
+        if (loggedin_user == None):
+                return redirect("login")   
+        else:                   
        
-        obj=Apply_Job.objects.filter(job_code=job_code).values()
-        session_uname=request.session.get('loggedinuser')
-        u=request.GET.get('subject4')
-        print("uu",u)     
-        my_string = ' '.join([str(element) for element in session_uname])
-        return render(request,"View_applications.html",{"obj":obj,"username":my_string }) 
+                obj=Apply_Job.objects.filter(job_code=job_code).values()
+                session_uname=request.session.get('loggedinuser')
+                u=request.GET.get('subject4')
+                print("uu",u)     
+                # my_string = ' '.join([str(element) for element in session_uname])
+                return render(request,"View_applications.html",{"obj":obj,"username": session_uname }) 
 
 class Respond_Application_View (View):
         def get(self,request,id):
-          
+          loggedin_user=request.session.get('loggedinuser')                    
+          if (loggedin_user == None):
+                return redirect("login")
+          else:
             obj=Apply_Job.objects.filter(id=id).values()
             print(obj[0]['photo'])
             session_uname=request.session.get('loggedinuser')
               
-            my_string = ' '.join([str(element) for element in session_uname])
+        #     my_string = ' '.join([str(element) for element in session_uname])
            
            
             status_update_by_get=Apply_Job.objects.filter(id=id).update(status="Viewed")
@@ -147,17 +169,24 @@ class Respond_Application_View (View):
             elif k =="REJECT":
                 status_update_by_button_click=Apply_Job.objects.filter(id=id).update(status="Rejected")
                
-            return render(request,"Status-Application.html",{"obj":obj,"username":my_string})
+            return render(request,"Status-Application.html",{"obj":obj,"username": session_uname})
        
 class Company_Profileview(View):
         def get(self,request):
-            loggedin_user=request.session.get('loggedinuser')
-            my_string = ' '.join([str(element) for element in  loggedin_user])
-            user_details=Compnay_Profile.objects.filter(username=my_string).values()
-            print(user_details[0]['company_name'])
-            return render(request,"company_profile.html",{"job":user_details,"username":my_string}) 
+                loggedin_user=request.session.get('loggedinuser')                    
+                if (loggedin_user == None):
+                        return redirect("login")
+                else:  
+                        session_uname=request.session.get('loggedinuser')
+                        #     my_string = ' '.join([str(element) for element in  loggedin_user])
+                        user_details=Compnay_Profile.objects.filter(username= session_uname).values()
+                        print(user_details[0]['company_name'])
+                        return render(request,"company_profile.html",{"job":user_details,"username": session_uname}) 
       
 def update_companyprofile(request, id):
+        loggedin_user=request.session.get('loggedinuser')                    
+        if (loggedin_user == None):
+                return redirect("login")                      
         obj =Compnay_Profile.objects.get(id=id)
        
         context = {
